[asan] make tid u32 instead of int
llvm-svn: 158074

+++ compiler-rt/lib/asan/asan_win.cc
@@ -219,7 +219,7 @@ u16 AtomicExchange(u16 *a, u16 new_val) {
   return new_val;
 }
 
-u16 AtomicExchange(u16 *a, u16 new_val) {
+u8 AtomicExchange(u8 *a, u8 new_val) {
   // FIXME: can we do this with a proper xchg intrinsic?
   u8 t = *a;
   *a = new_val;
